Implement hooks for admin workspace, profile menu and theme preference
